:recycle: refactor: #12 띄어쓰기 수정
